refactor(mqttClient): Refactor AsyncMqttClient to have connect method
